% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{Display_All_Palettes}
\alias{Display_All_Palettes}
\title{Main colour extractor
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}
\usage{
Display_All_Palettes()
}
\description{
Main colour extractor
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2
}
