% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_breaks.R
\name{format_nor}
\alias{format_nor}
\title{format_nor}
\usage{
format_nor(x, digits = 0, sig = NULL)
}
\arguments{
\item{x}{value}

\item{digits}{Number of digits after the decimal place (required)}

\item{sig}{Number of significant digits (optional)}
}
\description{
format_nor
}
