% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltable_css.R
\name{htmltable_css_cell}
\alias{htmltable_css_cell}
\title{CSS for htmltable cell}
\usage{
htmltable_css_cell(fontsize = 14, border_bottom = "1px dashed grey")
}
\arguments{
\item{fontsize}{fontsize}

\item{border_bottom}{Border bottom style}
}
\description{
CSS for htmltable cell
}
