% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltable_css.R
\docType{data}
\name{htmltable_css_table}
\alias{htmltable_css_table}
\title{CSS for htmltable table}
\format{
An object of class \code{character} of length 1.
}
\usage{
htmltable_css_table
}
\description{
CSS for htmltable table
}
\keyword{datasets}
