% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmltable_css.R
\name{htmltable_quick_style}
\alias{htmltable_quick_style}
\title{CSS for htmltable header}
\usage{
htmltable_quick_style(tab, widths = rep(round(100/ncol(tab)), ncol(tab)))
}
\arguments{
\item{tab}{tab}

\item{widths}{Vector}
}
\description{
CSS for htmltable header
}
