% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable.R
\name{huxtable_theme}
\alias{huxtable_theme}
\title{fhi_huxtable_theme}
\usage{
huxtable_theme(ht, position = "center")
}
\arguments{
\item{ht}{A huxtable}

\item{position}{Position}
}
\description{
fhi_huxtable_theme
}
