% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_mem.R
\name{make_influenza_threshold_chart}
\alias{make_influenza_threshold_chart}
\title{A treshold chart of the type used for influenza}
\usage{
make_influenza_threshold_chart(data, title, lang = "NB", weeks = c(40,
  20))
}
\arguments{
\item{data}{A data.table for a season of influenza}

\item{title}{the title for the chart}

\item{lang}{language of labels}

\item{weeks}{start and end week for the plot}
}
\description{
A treshold chart of the type used for influenza
}
