% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_a4.R
\name{save_a4}
\alias{save_a4}
\title{save_a4}
\usage{
save_a4(p, filename, landscape = T)
}
\arguments{
\item{p}{plot}

\item{filename}{filename}

\item{landscape}{landscape dimensions?}
}
\description{
save_a4
}
