% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_fill_fhi}
\alias{scale_fill_fhi}
\title{Main colour extractor
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}
\usage{
scale_fill_fhi(palette = "dis_primary", direction = 1, ...)
}
\arguments{
\item{palette}{a}

\item{direction}{a}

\item{...}{a}
}
\description{
Main colour extractor
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2
}
