% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_stepribbon.R
\name{stairstepn}
\alias{stairstepn}
\title{Internal function}
\usage{
stairstepn(data, direction = "hv", yvars = "y")
}
\arguments{
\item{data}{a}

\item{direction}{a}

\item{yvars}{a}
}
\description{
Internal function
}
