% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_stepribbon.R
\docType{data}
\name{StatStepribbon}
\alias{StatStepribbon}
\title{StatStepribbon}
\format{An object of class \code{stepribbon} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatStepribbon
}
\description{
StatStepribbon
}
\keyword{datasets}
