% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_fhi_lines}
\alias{theme_fhi_lines}
\title{theme_fhi_basic}
\usage{
theme_fhi_lines(base_size = 18, base_family = "")
}
\arguments{
\item{base_size}{a}

\item{base_family}{a}
}
\description{
theme_fhi_basic
}
