% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_breaks.R
\name{week_to_x}
\alias{week_to_x}
\title{week_to_x
Natural week to influenza season week. Influenza week 1 is natural week 30.}
\usage{
week_to_x(week)
}
\arguments{
\item{week}{Natural week in a year}
}
\description{
week_to_x
Natural week to influenza season week. Influenza week 1 is natural week 30.
}
