% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_breaks.R
\name{x_to_week}
\alias{x_to_week}
\title{x_to_week
Influenza season week to natural week. Influenza week 1 is natural week 30.}
\usage{
x_to_week(x)
}
\arguments{
\item{x}{Influenza week in a year}
}
\description{
x_to_week
Influenza season week to natural week. Influenza week 1 is natural week 30.
}
